chore: cleanup after merge
