+++ rust/src/operations/transaction/conflict_checker.rs
@@ -6,8 +6,9 @@ use std::io::{BufRead, BufReader, Cursor};
 use object_store::ObjectStore;
 use serde_json::{Map, Value};
 
-use super::{can_downgrade_to_snapshot_isolation, CommitInfo, IsolationLevel};
+use super::CommitInfo;
 use crate::action::{Action, Add, DeltaOperation, MetaData, Protocol, Remove};
+use crate::delta_config::IsolationLevel;
 use crate::operations::transaction::TransactionError;
 use crate::storage::{commit_uri_from_version, DeltaObjectStore, ObjectStoreRef};
 use crate::{
@@ -504,6 +505,56 @@ impl<'a> ConflictChecker<'a> {
     }
 }
 
+// implementation and comments adopted from
+// https://github.com/delta-io/delta/blob/1c18c1d972e37d314711b3a485e6fb7c98fce96d/core/src/main/scala/org/apache/spark/sql/delta/OptimisticTransaction.scala#L1268
+//
+// For no-data-change transactions such as OPTIMIZE/Auto Compaction/ZorderBY, we can
+// change the isolation level to SnapshotIsolation. SnapshotIsolation allows reduced conflict
+// detection by skipping the
+// [ConflictChecker::check_for_added_files_that_should_have_been_read_by_current_txn] check i.e.
+// don't worry about concurrent appends.
+//
+// We can also use SnapshotIsolation for empty transactions. e.g. consider a commit:
+// t0 - Initial state of table
+// t1 - Q1, Q2 starts
+// t2 - Q1 commits
+// t3 - Q2 is empty and wants to commit.
+// In this scenario, we can always allow Q2 to commit without worrying about new files
+// generated by Q1.
+//
+// The final order which satisfies both Serializability and WriteSerializability is: Q2, Q1
+// Note that Metadata only update transactions shouldn't be considered empty. If Q2 above has
+// a Metadata update (say schema change/identity column high watermark update), then Q2 can't
+// be moved above Q1 in the final SERIALIZABLE order. This is because if Q2 is moved above Q1,
+// then Q1 should see the updates from Q2 - which actually didn't happen.
+pub(super) fn can_downgrade_to_snapshot_isolation<'a>(
+    actions: impl IntoIterator<Item = &'a Action>,
+    operation: &DeltaOperation,
+    isolation_level: &IsolationLevel,
+) -> bool {
+    let mut data_changed = false;
+    let mut has_non_file_actions = false;
+    for action in actions {
+        match action {
+            Action::add(act) if act.data_change => data_changed = true,
+            Action::remove(rem) if rem.data_change => data_changed = true,
+            _ => has_non_file_actions = true,
+        }
+    }
+
+    if has_non_file_actions {
+        // if Non-file-actions are present (e.g. METADATA etc.), then don't downgrade the isolation
+        // level to SnapshotIsolation.
+        return false;
+    }
+
+    match isolation_level {
+        IsolationLevel::Serializable => !data_changed,
+        IsolationLevel::WriteSerializable => !data_changed && !operation.changes_data(),
+        IsolationLevel::SnapshotIsolation => false, // this case should never happen, since spanpshot isolation canot be configured on table
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::super::test_utils as tu;
@@ -527,6 +578,18 @@ mod tests {
             }
         });
         Some(data.to_string())
+    }
+
+    #[test]
+    fn test_can_downgrade_to_snapshot_isolation() {
+        let isolation = IsolationLevel::WriteSerializable;
+        let operation = DeltaOperation::Optimize {
+            predicate: None,
+            target_size: 0,
+        };
+        let add = tu::create_add_action("p", false, None);
+        let res = can_downgrade_to_snapshot_isolation(&[add], &operation, &isolation);
+        assert!(!res)
     }
 
     #[tokio::test]
